pyproject.toml
[project]
name = "bluecore-api"
version = "0.1.0"
description = "Blue Core API for managing BIBFRAME RDF data and workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]",
  "sqlalchemy",
  "pgvector",
  "psycopg2-binary>=2.9.10",
  "rdflib>=7.1.3",
  "bluecore-models>=0.4.3",
  "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
  "ruff",
  "pytest-mock-resources[docker]>=2.12.1",
  "pytest-mock>=3.14.0",
  "pytest-httpx>=0.35.0",
  "pytest-asyncio>=0.26.0",
  "pytest-ordering==0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bluecore_api"]
